chore: extract pixel utilities
