+++ src/components/Viewport.vue
@@ -68,7 +68,8 @@ import { useTemplateRef, computed, onMounted, onUnmounted } from 'vue';
 import { useStore } from '../stores';
 import { useService } from '../services';
 import { OVERLAY_STYLES, GRID_STROKE_COLOR } from '@/constants';
-import { rgbaCssU32, ensureCheckerboardPattern } from '../utils';
+import { rgbaCssU32 } from '../utils';
+import { ensureCheckerboardPattern } from '../utils/pixels.js';
 
 const { viewport: viewportStore, nodeTree, nodes, pixels: pixelStore, viewportEvent: viewportEvents } = useStore();
 const { overlay, toolSelection: toolSelectionService, viewport } = useService();
